On the request of tackat, change all floats to doubles.  We need the added precision when we are going to more detailed tiles.
An initial experiment of doing this automatically didn't come out too
well: All places except 3 disappeared, and the North Pole was placed
at the equator.  Instead I'm now doing it manually.  This commit
changes all floats in the classes Quaternion, GeoPoint and GeoPolygon,
but only as little as possible of the rest to make it compile.

From here on, I will take it step by step.


svn path=/trunk/KDE/kdeedu/marble/; revision=674069

+++ src/lib/GeoPoint.cpp
@@ -17,22 +17,22 @@
 #include <QtCore/QDebug>
 
 
-const float rad2int = 21600.0 / M_PI;
+const double rad2int = 21600.0 / M_PI;
 
 GeoPoint::GeoPoint(int _detail, int _lon, int _lat)
   : m_detail(_detail)
 {
-    m_q = Quaternion( (float)(_lon) / rad2int * 2,
-		      (float)(_lat) / rad2int * 2 );
+    m_q = Quaternion( (double)(_lon) / rad2int * 2,
+		      (double)(_lat) / rad2int * 2 );
 }
 
 GeoPoint::GeoPoint(int _lon, int _lat)
 {
-    m_q = Quaternion( (float)(_lon) / rad2int * 2,
-		      (float)(_lat) / rad2int * 2 );
+    m_q = Quaternion( (double)(_lon) / rad2int * 2,
+		      (double)(_lat) / rad2int * 2 );
 }
 
-GeoPoint::GeoPoint( float _lon, float _lat )
+GeoPoint::GeoPoint( double _lon, double _lat )
 {
     m_q = Quaternion( _lon, _lat );
 }
@@ -40,7 +40,7 @@ GeoPoint::GeoPoint( float _lon, float _lat )
 
 QString GeoPoint::toString()
 {
-    float lat, lon;
+    double lat, lon;
     geoCoordinates( lon, lat );
 
     QString nsstring = ( lat < 0 ) ? "N" : "S";  
@@ -50,13 +50,13 @@ QString GeoPoint::toString()
 
     int londeg = (int) lon;
     int lonmin = (int) ( 60 * (lon - londeg) );
-    int lonsec = (int) ( 3600 * (lon - londeg - ((float)(lonmin) / 60) ) );
+    int lonsec = (int) ( 3600 * (lon - londeg - ((double)(lonmin) / 60) ) );
 
     lat = fabs( lat * 180.0 / M_PI );
 
     int latdeg = (int) lat;
     int latmin = (int) ( 60 * (lat - latdeg) );
-    int latsec = (int) ( 3600 * (lat - latdeg - ((float)(latmin) / 60) ) );
+    int latsec = (int) ( 3600 * (lat - latdeg - ((double)(latmin) / 60) ) );
 
     return QString("%1\xb0 %2\' %3\" %4, %5\xb0 %6\' %7\" %8")
 	.arg(latdeg).arg(latmin).arg(latsec).arg(nsstring)
